Generic : Mega-hack to allow different passing organisation for variadic funcs.
Right now, the generic code assumes that it can determine the stack layout from
target callbacks that don't have complete information about the argument they
are processing (they only know the mode and the type - not it the argument is
named or not, for example).  This hack is not complete - but makes some
progress.

It looks like we really need the option for a target to override the layout
process - if it requires distinct stack layouts for different calling convs.

(cherry picked from commit 1090760b67eae80e5829f2dc4f7256ddd368ad22)