configure, Darwin : Adjust RPATH_ENVVAR for embedded run-paths.
Darwin versions later than Darwin14 cannot use DYLD_LIBRARY_PATH to
provide the run-path for stage 1 target and stage 2+ configure, since
system tools like /bin/sh remove the environment variable.  If we want
to access this info we will have to do so explicitly; so provide a
separate (non DYLD_) environment variable for this.  We default to an
embedded run-path for these OS versions.
Embedded run-paths are optional (but recommended) on Darwin9 to 14.

Signed-off-by: Kirill A. Korinsky <kirill@korins.ky>